lint clean.
git-svn-id: file:///svn/dnssec-trigger/trunk@472 14dc9c71-5cc2-e011-b339-0019d10b89f4

+++ riggerd/probe.c
@@ -284,8 +284,8 @@ check_ssl(struct outq* outq)
 			size_t blen = sizeof(buf);
 			unsigned int i;
 			for(i=0; i<len; i++) {
-				size_t a = snprintf(at, blen, "%2.2X%s",
-					(int)hash[i], (i==len-1)?"":":");
+				int a = snprintf(at, blen, "%2.2X%s",
+				    (unsigned int)hash[i], (i==len-1)?"":":");
 				at += a; 
 				blen -= a;
 			}
